uploaded/tai-GaaS-bugfix-sankeagrawal-fe-be-issues/Legacy_Demo/requirements.txt
chromadb==0.4.10
llm-guard==0.3.0
langchain==0.0.352
langchain-experimental==0.0.4
openai==0.27.9
openpyxl==3.1.2
Pillow==10.0.1
pypdf==3.17.0
python-dotenv==1.0.0
seaborn==0.12.2
seleniumbase==4.18.8
sentence-transformers==2.2.2
st_pages==0.4.4
streamlit==1.27.2
streamlit_extras==0.3.1
tabulate==0.9.0
tiktoken==0.5.1
toml==0.10.2
python-docx==0.8.11
lxml==4.9.3
streamlit-tags==1.2.8
ai-server-sdk
